+++ src/main/java/com/financemanager/service/impl/DefaultIncomeService.java
@@ -1,6 +1,7 @@
 package com.financemanager.service.impl;
 
 import com.financemanager.dto.IncomeDTO;
+import com.financemanager.entity.CustomUserDetails;
 import com.financemanager.entity.utils.DatePart;
 import com.financemanager.exception.ResourceNotFoundException;
 import com.financemanager.mapper.IncomeMapper;
@@ -13,6 +14,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -46,6 +49,7 @@ public class DefaultIncomeService implements IncomeService {
 
     @Override
     public IncomeDTO update(IncomeDTO incomeDTO, Long id) {
+        checkPermission(incomeDTO);
         incomeRepository.findById(id).orElseThrow(
                 () -> new ResourceNotFoundException(String.format(incomeNotFoundError, id)));
         return incomeMapper.toIncomeDTO(incomeRepository.save(incomeMapper.toIncome(incomeDTO)));
@@ -55,14 +59,23 @@ public class DefaultIncomeService implements IncomeService {
     public void delete(Long id) {
         IncomeDTO incomeDTO = incomeRepository.findById(id).map(incomeMapper::toIncomeDTO).orElseThrow(
                 () -> new ResourceNotFoundException(String.format(incomeNotFoundError, id)));
+        checkPermission(incomeDTO);
         incomeRepository.deleteById(incomeDTO.getId());
     }
 
     @Override
     public IncomeDTO findById(Long id) {
-        IncomeDTO income = incomeRepository.findById(id).map(incomeMapper::toIncomeDTO).orElseThrow(
+        IncomeDTO incomeDTO = incomeRepository.findById(id).map(incomeMapper::toIncomeDTO).orElseThrow(
                 () -> new ResourceNotFoundException(String.format(incomeNotFoundError, id)));
-        return income;
+        checkPermission(incomeDTO);
+        return incomeDTO;
+    }
+    
+    private void checkPermission(IncomeDTO incomeDTO){
+        CustomUserDetails user = (CustomUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(incomeDTO.getUserId() != user.getId() && !user.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_ADMIN"))) {
+           throw new AccessDeniedException("You don't have permission for this action"); 
+        }
     }
 
 }
